requirements.txt
future
emoji==1.7.0
lxml
wget
requests 
SQLAlchemy==1.4.40
python-telegram-bot==13.11
gtts
webcolors
beautifulsoup4
pytimeparse
psycopg2-binary
feedparser
faker
hachoir
pykeyboard
pynewtonmath
aiofiles
ffmpeg-python
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp
Pillow
CurrencyConverter
googletrans
google-trans-new
jikanpy
speedtest-cli
coffeehouse
regex
bleach
wikipedia
markdown2
telethon
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
pygithub>=1.55
pyrogram>=2.0.35
TgCrypto
PyDictionary
git+https://github.com/Sorrybabe/downloader
youtube_search_python==1.4.9
EmojiCaptcha==0.1.2
captcha==0.4 
youtube_search
lyrics_extractor
asyncio
cloudscraper
dateparser
pymongo==3.12.1
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4==0.0.1
redis
pySmartDL
bing_image_downloader
search_engine_parser
pytz
jikanpy
tswift
python-arq
envparse
#profanity (Credits to Julia)
better_profanity
textblob 
nudepy
nekos.py
motor
gpytranslate
fontTools
opencv-python-headless
odmantic
tzlocal
pendulum
fuzzysearch
img2pdf
multicolorcaptcha
git+https://github.com/python-telegram-bot/ptbcontrib.git
loguru
pyYAML>=5.1.2
python-barcode==0.14.0
